add conda_build_config.yaml project-wide config
In several PR's, I have been saying that those PR's depend on a conda_build_config.yaml that I did not include.  This is that conda_build_config.yaml file.  This file establishes the overall version space for all builds.

The key implementation detail here is that rendering is done with the current working directory being the root of this repo.  Thus, this conda_build_config.yaml takes effect for all recipes here.  It can be overridden (reduced, expanded, or changed) by conda_build_config.yaml files that live alongside meta.yaml files in any of the recipes.