python.app: Force creation of the lib symlink and ignore any failure
This is so that force reinstalls from the installers do not fail in the
post-link script.

Although ignoring failure may not sound like a good idea, force installing
Miniconda is sometimes the only recourse people have so I would rather it
not fail on something more minor like failing to make this symlink.

Fixes: https://github.com/ContinuumIO/anaconda-issues/issues/6915