(WIP) Added cctools-ld64-feedstock
I needed to debug ld64 but having to build all of clang to do that is not good.
Also, extracting this will be necessary to keep our clang/llvm recipe clean.

WIP because libxar.a is not fully static. It seems libxml2 links to ICU instead
of iconv. I need to make a libxml2-bootstrap which links to iconv instead. ICU
is too large to statically link to ld64. Then I need to make a xar-bootstrap
which uses this libxml2-bootstrap.